python-service/requirements.txt
# VeriResume Python AI Service Requirements

# Core Framework
flask==3.0.0
flask-cors==4.0.0

# Resume Parsing
pdfplumber>=0.11.0
python-docx==1.1.0
spacy==3.7.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl

# AI Analysis
openai==1.6.1
google-generativeai==0.3.2
groq==0.4.1

# NLP & Text Processing
sentence-transformers==2.2.2
scikit-learn==1.3.2
nltk==3.8.1

# Job Scraping
selenium==4.16.0
beautifulsoup4==4.12.2
requests==2.31.0

# Date/Time
python-dateutil==2.8.2

# Environment
python-dotenv==1.0.0

# Utilities
Pillow==10.1.0
pandas==2.1.4
numpy==1.26.2
